# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Hijrah App API",
        "description": "Complete API documentation for Hijrah App - A comprehensive assessment and program management system",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/program/{programId}/step/{stepId}/start": {
            "post": {
                "tags": [
                    "Programs"
                ],
                "summary": "Start a step",
                "description": "Start a step for the authenticated user. Creates step progress tracking with 'in_progress' status.",
                "operationId": "startStep",
                "parameters": [
                    {
                        "name": "programId",
                        "in": "path",
                        "description": "Program ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "stepId",
                        "in": "path",
                        "description": "Step ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Step started successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Step started successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error starting step",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error starting step"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/program/{programId}/step/{stepId}/complete": {
            "post": {
                "tags": [
                    "Programs"
                ],
                "summary": "Complete a step",
                "description": "Complete a step with type-specific data. Journal steps require 'thought', Quiz steps require 'score', Challenge steps require 'challenges_done' and 'percentage'.",
                "operationId": "completeStep",
                "parameters": [
                    {
                        "name": "programId",
                        "in": "path",
                        "description": "Program ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "stepId",
                        "in": "path",
                        "description": "Step ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "description": "Step completion data (varies by step type)",
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "$ref": "#/components/schemas/JournalStepCompletionRequest"
                                    },
                                    {
                                        "$ref": "#/components/schemas/QuizStepCompletionRequest"
                                    },
                                    {
                                        "$ref": "#/components/schemas/ChallengeStepCompletionRequest"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Step completed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Step completed successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error completing step",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error completing step"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Step not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Step not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation failed"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "thought": [
                                                    "The thought field is required."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/program/{programId}/step/{stepId}/challenge-progress": {
            "post": {
                "tags": [
                    "Programs"
                ],
                "summary": "Update challenge progress",
                "description": "Update the progress of challenge-type steps by tracking individual challenges completed. Automatically calculates percentage based on total challenges and sets step to 'in_progress' status.",
                "operationId": "updateChallengeProgress",
                "parameters": [
                    {
                        "name": "programId",
                        "in": "path",
                        "description": "Program ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "stepId",
                        "in": "path",
                        "description": "Step ID (must be a challenge-type step)",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "description": "Challenge progress data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ChallengeProgressRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Challenge progress updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ChallengeProgressResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid step type or other error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid step type for challenge progress"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation failed"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "challenges_done": [
                                                    "The challenges done field is required."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "JournalStepCompletionRequest": {
                "required": [
                    "thought"
                ],
                "properties": {
                    "thought": {
                        "description": "Reflection text for journal step completion",
                        "type": "string",
                        "maxLength": 2000,
                        "example": "Today I reflected on my spiritual journey and realized the importance of consistency in my daily practices."
                    }
                },
                "type": "object"
            },
            "QuizStepCompletionRequest": {
                "required": [
                    "score"
                ],
                "properties": {
                    "score": {
                        "description": "Quiz score between 0-100",
                        "type": "integer",
                        "maximum": 100,
                        "minimum": 0,
                        "example": 85
                    }
                },
                "type": "object"
            },
            "ChallengeStepCompletionRequest": {
                "required": [
                    "challenges_done",
                    "percentage"
                ],
                "properties": {
                    "challenges_done": {
                        "description": "Number of challenges completed",
                        "type": "integer",
                        "minimum": 0,
                        "example": 7
                    },
                    "percentage": {
                        "description": "Completion percentage",
                        "type": "number",
                        "format": "float",
                        "maximum": 100,
                        "minimum": 0,
                        "example": 87.5
                    }
                },
                "type": "object"
            },
            "ChallengeProgressRequest": {
                "required": [
                    "challenges_done"
                ],
                "properties": {
                    "challenges_done": {
                        "description": "Number of challenges completed so far",
                        "type": "integer",
                        "minimum": 0,
                        "example": 5
                    }
                },
                "type": "object"
            },
            "ChallengeProgressResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Challenge progress updated successfully"
                    },
                    "data": {
                        "properties": {
                            "challenges_done": {
                                "type": "integer",
                                "example": 5
                            },
                            "total_challenges": {
                                "type": "integer",
                                "example": 10
                            },
                            "percentage": {
                                "type": "number",
                                "format": "float",
                                "example": 50
                            },
                            "status": {
                                "type": "string",
                                "enum": [
                                    "not_started",
                                    "in_progress",
                                    "completed"
                                ],
                                "example": "in_progress"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Use a bearer token to access this endpoint",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}